requirements.txt
streamlit==1.37.1
pandas==2.2.2
PyMuPDF==1.24.9
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.33
langchain-text-splitters==0.2.2
python-dotenv==1.0.1
openai==1.41.0
